backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Cors==3.0.10
Flask-JWT-Extended==4.6.0
python-decouple==3.8
pyotp==2.8.0
pytest==7.4.2
pytest-cov==4.1.0
freezegun==1.2.2
psycopg2-binary==2.9.9
requests==2.32.5
boto3==1.40.16
python-magic-bin==0.4.14 #to check
docusign-esign==5.2.0
moto==5.1.11
